client: new package
Moving api_client.go and api_client_test.go from the
github.com/dotcloud/docker path.
